+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.1.1] - 2023-04-27 (Una mascota trapellota)
+
+- Add tasks to migrate from S3 to :local storage
+- Better internal API
+
 ## [0.1.0] - 2023-04-18 (Pets i espetecs)
 
 - Initial release